backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.8.6
protobuf>=5.29.5
supabase==2.0.3
requests==2.31.0
httpcore>=1.0.9
httpx>=0.28.1
pdfplumber==0.11.0

